+++ Android/app/src/main/java/com/lawgicalai/bubbychat/data/di/utils/Extensions.kt
@@ -1,4 +1,5 @@
 package com.lawgicalai.bubbychat.data.di.utils
 
 fun String?.isJsonObject(): Boolean = this?.startsWith("{") == true && this.endsWith("}")
+
 fun String?.isJsonArray(): Boolean = this?.startsWith("[") == true && this.endsWith("]")
