+++ Android/app/src/main/java/com/lawgicalai/bubbychat/data/usecase/token/SetTokenUseCaseImpl.kt
@@ -4,10 +4,12 @@ import com.lawgicalai.bubbychat.data.datastore.UserDataStore
 import com.lawgicalai.bubbychat.domain.usecase.SetTokenUseCase
 import javax.inject.Inject
 
-class SetTokenUseCaseImpl @Inject constructor(
-    private val userDataStore: UserDataStore
-): SetTokenUseCase {
-    override suspend fun invoke(token: String) {
-        userDataStore.setToken(token)
+class SetTokenUseCaseImpl
+    @Inject
+    constructor(
+        private val userDataStore: UserDataStore,
+    ) : SetTokenUseCase {
+        override suspend fun invoke(token: String) {
+            userDataStore.setToken(token)
+        }
     }
-}
