--- Android/app/src/main/java/com/lawgicalai/bubbychat/domain/model/Dummy.kt
@@ -1,2 +1,0 @@
-package com.lawgicalai.bubbychat.domain.model
-
