+++ Android/app/src/main/java/com/lawgicalai/bubbychat/domain/model/Persona.kt
@@ -1,0 +1,7 @@
+package com.lawgicalai.bubbychat.domain.model
+
+data class Persona(
+    val type: String,
+    val name: String,
+    val description: String,
+)
